Still compilerlistener
This takes in a function like fun(){return 1;} with a few caveats. Line 122 I use symtab.put(id, 0); because otherwise we get the "function not defined" error. We should eventually find out what the symbol table is useful for, besides making lots of things not work. Also, if fun is return 0; instead of return 1; or return 2; it will do that loop thing where the last RET in main takes it to the top of the file instead of line 3 for the ADDSP and STOP. I also tried it with return getint() and i believe it is the same problem, constantly taking new user input. stepping through it, if you enter 0 for get int it will return to the top of the file, but if you enter something else like 1, the line tracker disappears, not returning to anywhere, just gone and ends the program. Not sure because when not debugging it continues to accept.