Still compromiseliaatener
Updated for the case in which there are two if statements in the program that aren't nested. Should now offset the numbering by a factor of ten. I could get around all of this by using the e ID numbering, but I have no idea what that is. One issue with this, it doesn't like return putint((10?20:30)+(10?20:30));. We aren't technically supposed to be able to parse expr op expr, only term op expr, and an if expression isn't a term, but that feels like an oversight? Not sure.